make a scrape button component. working on getting a promise tto grab the image and then click on it and open up and grab the link to the biggger image
