+++ client/src/components/Button.js
@@ -1,10 +1,4 @@
 import React from "react";
-import styled from "styled-components";
-
-const Input = styled.input`
-border-radius: 3px;
-margin: 0.5em;
-`;
 
 const Button = props => {
